requirements.txt
certifi
Click
Flask
gunicorn
itsdangerous
Jinja2
MarkupSafe
Werkzeug
numpy
scipy
matplotlib
pandas
tpot
cloudpickle
azure-cognitiveservices-speech
pydub